Support for tagging sophisticated TOCs such as mind maps in the manual
\etoctagstartTOCcontents
\etocthetocitembegintag
\etoctagReference
\etoctagLbl
\etocthetocitemendtag
\etoctagfinishTOCcontents

This induced a number of internal refactoring.

Attention that \Etoc@lxyz is now done before the contentsline tagging
begin code of upstream kernel.

Also add \etochyperlink as \protected wrapper of \hyperlink, and use it
in \etocthelinkedname, \etocthelinkednumber and \etocthelinkedpage. a
They thus can be used in an \edef, which facilitates the sophisticated
things such as the TOC as a tree, TOC as a molecule, TOC as mind map in
the user manual.

But in some cases \etocthename may still contain things that do not like
\edef.

	modified:   ChangeLog.md
	modified:   etoc.sty
	modified:   etoctagging.md